// src/webparts/polRssGallery/components/RssFeed.module.scss.d.ts
export interface IRssFeedModuleScss {
  webpart: string;
  bannerContainer: string;
  bannerSlide: string;
  bannerImage: string;
  bannerImagePlaceholder: string;
  bannerCaption: string;
  bannerTitle: string;
  cardGrid: string;
  card: string;
  cardImage: string;
  cardImagePlaceholder: string;
  cardBody: string;
  cardTitle: string;
  cardDescription: string;
  list: string;
  listItem: string;
  listItemThumbnail: string;
  listItemTitle: string;
  error: string;
  noItems: string;
  placeholder: string;
}

declare const styles: IRssFeedModuleScss;
export default styles;